Update description of ComServer element
Packaged COM registrations in the manifest can only get Activate As Package behavior, but the current description makes it sound like that is just the default and that developers can opt into Activate As Activator behavior.  I modified the description to make it clear that Activate As Package is the only option while still explaining how it diverges from other activation behaviors.

Also, the description suggests using the ApplicationId value on the AppID registry key to select an Applications/Application element for the ComServer.  While this is correct, the ApplicationId value is new in the Anniversary Update, so few COM servers will have it in their registration.  Given that Store policy only allows multi-application packages with special approval, I think it is better to avoid confusing developers with details that rarely apply to them.  The description of when and how to use the ApplicationId value is better suited for the AppID key's MSDN page: https://msdn.microsoft.com/library/windows/desktop/ms682359.aspx.