Fix winrt::com_ptr put/put_void to describe the updated behavior
in the initial cppwinrt design these methods did not clear the value as an optimization, backed up with an assert. but since many clients don't build with asserts on, it is very easy for bugs to creep in. so the design was changed to this safer behavior, exhibited by other com smart pointer implementations.