Clarify winrt::guid_of documentation
Based on discussion in https://github.com/microsoft/cppwinrt/issues/1494 where @kennykerr stated the following:
The guid_of function returns the identifier of the WinRT type. For COM and WinRT interfaces, this is the identifier of the interface. For WinRT classes, this is required to be the same as that of the default interface of the class.
It is only if and when guid_v is not specialized that it will fallback to using __uuidof where applicable.